[MAT-180] Attempt to replace all sizes with size_t
This involves those values not directly touched by the JVM.

Because all sizes of things are replaced with size_t, it is not
possible for negative sizes to be used any more - the code that
checks for this and the tests of this check are removed.

However, the fields holding sizes are signed in Java. To ensure
that the passing of sizes from Java to C++ is safe, we replace the
getIntFromVoidJMethod with getSizeTFromVoidJMethod that checks
that the received value is non-negative, before casting it to
size_t.

This should leave the remaining ints (at least in librdag) all
safe to convert to int4.

+++ include/equals.hh
@@ -18,27 +18,27 @@ extern const real8 FuzzyEquals_default_maxrelerror;
  * Bitwise equals, equals via memcmp()
  */
 template <typename T>
-bool ArrayBitEquals(T * arr1, T * arr2, int count);
+bool ArrayBitEquals(T * arr1, T * arr2, size_t count);
 
 extern template bool
-ArrayBitEquals(real8 * arr1, real8 * arr2, int count);
+ArrayBitEquals(real8 * arr1, real8 * arr2, size_t count);
 
 extern template bool
-ArrayBitEquals(complex16 * arr1, complex16 * arr2, int count);
+ArrayBitEquals(complex16 * arr1, complex16 * arr2, size_t count);
 
 extern template bool
-ArrayBitEquals(int * arr1, int * arr2, int count);
+ArrayBitEquals(int * arr1, int * arr2, size_t count);
 
 /**
  * Fuzzy equals, equals within tolerance specified
  */
 template<typename T>
-bool ArrayFuzzyEquals(T * arr1, T * arr2, int count, real8 maxabserror = FuzzyEquals_default_maxabserror, real8 maxrelerror = FuzzyEquals_default_maxrelerror);
+bool ArrayFuzzyEquals(T * arr1, T * arr2, size_t count, real8 maxabserror = FuzzyEquals_default_maxabserror, real8 maxrelerror = FuzzyEquals_default_maxrelerror);
 
 extern template bool
-ArrayFuzzyEquals(real8 * arr1, real8 * arr2, int count, real8 maxabserror, real8 maxrelerror);
+ArrayFuzzyEquals(real8 * arr1, real8 * arr2, size_t count, real8 maxabserror, real8 maxrelerror);
 extern template bool
-ArrayFuzzyEquals(complex16 * arr1, complex16 * arr2, int count, real8 maxabserror, real8 maxrelerror);
+ArrayFuzzyEquals(complex16 * arr1, complex16 * arr2, size_t count, real8 maxabserror, real8 maxrelerror);
 
 bool SingleValueFuzzyEquals(real8 val1, real8 val2, real8 maxabserror = FuzzyEquals_default_maxabserror, real8 maxrelerror = FuzzyEquals_default_maxrelerror);
 bool SingleValueFuzzyEquals(complex16 val1, complex16 val2, real8 maxabserror = FuzzyEquals_default_maxabserror, real8 maxrelerror = FuzzyEquals_default_maxrelerror);
